+++ src/index.ts
@@ -605,6 +605,7 @@ export default class Elysia<
 						modules,
 						validator: cloned.cookie,
 						defaultConfig: this.config.cookie,
+						normalize,
 						config: cloned.cookie?.config ?? {},
 						dynamic,
 						models,
@@ -699,6 +700,7 @@ export default class Elysia<
 									modules,
 									dynamic,
 									models,
+									normalize,
 									additionalProperties: !normalize,
 									coerce: true,
 									additionalCoerce:
@@ -719,6 +721,7 @@ export default class Elysia<
 									modules,
 									dynamic,
 									models,
+									normalize,
 									coerce: true,
 									additionalCoerce:
 										stringToStructureCoercions(),
@@ -738,6 +741,7 @@ export default class Elysia<
 									modules,
 									dynamic,
 									models,
+									normalize,
 									coerce: true,
 									additionalCoerce:
 										stringToStructureCoercions(),