+++ src/schema.ts
@@ -974,11 +974,13 @@ export const getSchemaValidator = <T extends TSchema | string | undefined>(
 			console.warn(
 				'Failed to create exactMirror. Please report the following code to https://github.com/elysiajs/elysia/issues'
 			)
-			console.warn(schema)
+			console.dir(schema, {
+				depth: null
+			})
 
 			compiled.Clean = createCleaner(schema)
 		}
-	} else compiled.Clean = createCleaner(schema)
+	} else if (normalize === 'typebox') compiled.Clean = createCleaner(schema)
 
 	compiled.parse = (v) => {
 		try {
@@ -1277,6 +1279,7 @@ export const getCookieValidator = ({
 	defaultConfig = {},
 	config,
 	dynamic,
+	normalize = false,
 	models,
 	validators,
 	sanitize
@@ -1286,6 +1289,7 @@ export const getCookieValidator = ({
 	defaultConfig: CookieOptions | undefined
 	config: CookieOptions
 	dynamic: boolean
+	normalize: ElysiaConfig<''>['normalize'] | undefined
 	models: Record<string, TSchema> | undefined
 	validators?: InputSchema['cookie'][]
 	sanitize?: () => ExactMirrorInstruction['sanitize']
@@ -1294,6 +1298,7 @@ export const getCookieValidator = ({
 		modules,
 		dynamic,
 		models,
+		normalize,
 		additionalProperties: true,
 		coerce: true,
 		additionalCoerce: stringToStructureCoercions(),